Make waitUntil = "load" the default for new items
It's a lot faster and seems to work for many sites, so probably
best to try with that first.

Note that everything that's currently
`waitUntil = "domcontentloaded"` hasn't been tested using `"load"`
instead, so that might very well work for them as well; I limited
my testing to the ones that logged as "couldn't fetch" in one
build.